Thanks, Pahimar.
Since we want to do blending, it seems appropriate to also set the blend
function -- to the same one Pahimar uses for his Minium/Philosopher
Stone in-world transmutation preview. It's pretty much the right look,
after all.

This should prevent any more strange flickering when rendering the held
block preview, and be somewhat more robust against leaky render context
from other mods.

Rendering is so evil, when it goes to Hell, the Devil will be retiring.